Consolidate two generators into one for better user experience
