+++ 90stratis/module-setup.sh
@@ -19,10 +19,12 @@ installkernel() {
 
 # called by dracut
 install() {
+    # Stratis dependencies
     inst_multiple stratis-min thin_check thin_repair mkfs.xfs xfs_admin xfs_growfs
-    inst_multiple $systemdutildir/system-generators/stratis-rootfs-prompt-generator \
-	    $systemdutildir/system-generators/stratis-setup-generator \
-	    $systemdutildir/stratis-key-set \
+    # Dracut dependencies
+    inst_multiple $systemdutildir/system-generators/stratis-setup-generator \
+	    $systemdutildir/stratis-open-rootfs \
+	    $systemdutildir/stratis-rootfs-setup \
 	    plymouth
 
     inst_rules "$moddir/11-stratisd.rules"
