+++ Cargo.toml
@@ -22,11 +22,6 @@ name = "stratis-min"
 required-features = ["extras"]
 
 [[bin]]
-name = "stratis-rootfs-prompt-generator"
-path = "src/bin/generators/stratis-rootfs-prompt-generator.rs"
-required-features = ["extras"]
-
-[[bin]]
 name = "stratis-setup-generator"
 path = "src/bin/generators/stratis-setup-generator.rs"
 required-features = ["extras"]
