+++ src/bin/generators/stratis-setup-generator.rs
@@ -4,27 +4,34 @@
 
 use std::path::PathBuf;
 
+use uuid::Uuid;
+
 mod lib;
 
-fn unit_template(uuids: Vec<PathBuf>) -> String {
+fn unit_template(uuids: Vec<PathBuf>, pool_uuid: Uuid) -> String {
     let devices: Vec<_> = uuids
         .into_iter()
         .map(|uuid_path| lib::encode_path_to_device_unit(&uuid_path))
         .collect();
     format!(
         r"[Unit]
-Description=setup for Stratis root filesystem
-Requires=stratis-rootfs-prompt.service
-After=stratis-rootfs-prompt.service
+Description=prompt for root filesystem password
+DefaultDependencies=no
+Conflicts=shutdown.target
+After=paths.target plymouth-start.service
+Before=initrd.target
 {}
 {}
 
 [Service]
 Type=oneshot
-ExecStart=/usr/bin/stratis-min pool setup
+Environment='STRATIS_ROOTFS_UUID={}'
+ExecStart=/usr/lib/systemd/stratis-rootfs-setup
+RemainAfterExit=yes
 ",
         format!("Requires={}", devices.join(" ")),
         format!("After={}", devices.join(" ")),
+        pool_uuid,
     )
 }
 
@@ -36,9 +43,16 @@ fn main() -> Result<(), String> {
         .get("stratis.rootfs.uuid_paths")
         .and_then(|opt_s| opt_s.as_ref().map(|s| s.to_string()))
         .ok_or_else(|| "Missing kernel command line parameter stratis.rootfs.uuids".to_string())?;
+    let pool_uuid = kernel_cmdline
+        .get("stratis.rootfs.pool_uuid")
+        .and_then(|opt_s| opt_s.as_ref().map(|s| s.to_string()))
+        .ok_or_else(|| {
+            "Missing kernel command line parameter stratis.rootfs.pool_uuid".to_string()
+        })?;
     let parsed_rootfs_uuid_paths: Vec<_> =
         rootfs_uuid_paths.split(',').map(PathBuf::from).collect();
-    let file_contents = unit_template(parsed_rootfs_uuid_paths);
+    let parsed_pool_uuid = Uuid::parse_str(&pool_uuid).map_err(|e| e.to_string())?;
+    let file_contents = unit_template(parsed_rootfs_uuid_paths, parsed_pool_uuid);
     let mut path = PathBuf::from(early_dir);
     path.push("stratis-setup.service");
     lib::write_unit_file(&path, file_contents).map_err(|e| e.to_string())?;
